Add command to calculate and get nonce
With this change there are also some changes to the CRC calculation,
since the previous way of doing calculcations were done on the struct
instead of the serialized packet.

Note that this commit uses hard coded indata to the nonce function, that
shall be replaced later on with user provided data. Likewise, it shall
later on also support passthrough mode.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>