Personalize: add running CRC calculation
Add a running CRC when programming data and OTP, this CRC should be
provided to the Lock command to ensure that data stored on the ATSHA204A
is the same as the stored data.

I failed to get that to work, even though I double checked by copy each
data slot and OTP to a separate buffer where I computed the CRC. That
CRC and the running CRC ended up exactly the same. So this needs to be
investigated and fixed in the future so that CRC check is used when
doing personalization.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>