Use the count field to locate the position of the CRC
There are cases where a Status/Error is returned, yet the number of bytes
received exceeds 4. Use the count field to determine whether a packet is
an Error/Status, as well as to locate the position of the CRC.

Check that the value contained in the count field is not smaller than the
number of bytes received.

Perform sanity checks before reading the incoming data.