Move tests into separate build
Move tests into own directory and create a new build target.
Replace main with stub.

Since we usually cross-compile, we don't use CTest. Instead,
create a new build target to generate a separate executable
to run on the target board.

Tests are not compiled by default. To build, run:

  make tests

Notice that the target is called "tests" and not "test" which
is the common case. This is because CMake reserves "test",
presumably for CTest.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>