Update Configuration zone access functions for atecc508a
In atsha204a, the Configuration zone is written in words, ie
each write is 4 Bytes long. In atecc508a, the Configuration
zone is written in blocks, ie 32 Bytes at a time. Update
s96_write_config() and its documentation to reflect the above.

Reading from the Configuration zone is performed in multiples of
word size in both devices. Update the documentation accordingly.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>